fix most of the warnings enabled by masochists
most of these people like to inflict pain on themselfs (i.e., warnings
in their own code), but they usually don't like if pain is inflicted
on them by others (i.e., warnings produced by external code which they
use). This patch should make these kinds of people happy. I'm not
really sure if the code is easier to understand with this, but at
least clang does not complain for most of the warnings of
"-Weverything" anymore.

+++ opm/material/fluidmatrixinteractions/PiecewiseLinearTwoPhaseMaterialParams.hpp
@@ -26,8 +26,8 @@
 #define OPM_PIECEWISE_LINEAR_TWO_PHASE_MATERIAL_PARAMS_HPP
 
 #include <vector>
-
 #include <cassert>
+#include <cstddef>
 
 namespace Opm {
 /*!
@@ -113,7 +113,7 @@ public:
     {
         assert(SwValues.size() == values.size());
 
-        int n = SwValues.size();
+        size_t n = SwValues.size();
         SwPcwnSamples_.resize(n);
         pcwnSamples_.resize(n);
 
@@ -141,7 +141,7 @@ public:
     {
         assert(SwValues.size() == values.size());
 
-        int n = SwValues.size();
+        size_t n = SwValues.size();
         SwKrwSamples_.resize(n);
         krwSamples_.resize(n);
 
@@ -169,7 +169,7 @@ public:
     {
         assert(SwValues.size() == values.size());
 
-        int n = SwValues.size();
+        size_t n = SwValues.size();
         SwKrnSamples_.resize(n);
         krnSamples_.resize(n);
 
@@ -195,7 +195,7 @@ private:
                  origSampleIdx < swValues.size() / 2;
                  ++ origSampleIdx)
             {
-                unsigned newSampleIdx = swValues.size() - origSampleIdx - 1;
+                size_t newSampleIdx = swValues.size() - origSampleIdx - 1;
 
                 std::swap(swValues[origSampleIdx], swValues[newSampleIdx]);
                 std::swap(values[origSampleIdx], values[newSampleIdx]);
